+++ controllers/softwarefactory_controller.go
@@ -9,6 +9,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
 	"sort"
 	"strconv"
 	"time"
@@ -22,7 +23,6 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
-	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	k8s_errors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -58,7 +58,6 @@ type SoftwareFactoryReconciler struct {
 //+kubebuilder:rbac:groups=*,resources=jobs/status;pods/status;services/status;statefulsets/status;deployments/status;configmaps/status;secrets/status;persistentvolumeclaims/status;serviceaccounts/status;roles/status,verbs=get
 //+kubebuilder:rbac:groups=cert-manager.io,resources=*,verbs=get;list;watch;create;update;patch;delete
 //+kubebuilder:rbac:groups=monitoring.coreos.com,resources=servicemonitors;podmonitors;prometheusrules,verbs=get;list;watch;create;update;patch;delete
-//+kubebuilder:rbac:groups=apiextensions.k8s.io,resources=customresourcedefinitions,verbs=list
 
 type SFController struct {
 	SFUtilContext
@@ -382,20 +381,14 @@ func (r *SFController) DebugService(debugService string) {
 	}
 }
 
-func checkOpenShift(ctx context.Context, client client.Client) bool {
-	crds := &apiextensions.CustomResourceDefinitionList{}
-	if err := client.List(ctx, crds); err != nil {
-		utils.LogI(fmt.Sprintf("Couldn't list crds, assume this is not openshift, %s", err))
-		return false
+func CheckOpenShift() bool {
+	env := os.Getenv("OPENSHIFT_USER")
+	openshiftUser, err := strconv.ParseBool(env)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "The OPENSHIFT_USER environment variable must be set to true/false, it was set to '%s'\n", env)
+		os.Exit(1)
 	}
-	for _, crd := range crds.Items {
-		if crd.Spec.Names.Kind == "Route" {
-			utils.LogI("API has Route, assume it is OpenShift")
-			return true
-		}
-	}
-	utils.LogI("API has no Route, assume it is vanilla Kubernetes")
-	return false
+	return openshiftUser
 }
 
 func (r *SoftwareFactoryReconciler) mkSFController(
@@ -413,7 +406,7 @@ func (r *SoftwareFactoryReconciler) mkSFController(
 			standalone: standalone,
 		},
 		cr:          cr,
-		isOpenShift: checkOpenShift(ctx, r.Client),
+		isOpenShift: CheckOpenShift(),
 	}
 }