+++ doc/reference/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 ## [in development]
 
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+
+## [v0.0.54] - 2025-02-26
+
+### Added
 
 - A new OPENSHIFT_USER environment variable is required to configure the operator to setup OpenShift attributes.
 
@@ -13,8 +21,6 @@ All notable changes to this project will be documented in this file.
 - storage: topolvm-provisioner is no longer the default storageClassName; if storageClassName parameters
   are not set, use the cluster's default storage class for persistent volumes.
 
-### Deprecated
-### Removed
 ### Fixed
 
 - OpenShift detection logic required the permission to list crds, this has been changed to use a new OPENSHIFT_USER environment variable.